requirements.txt
itchat-uos==1.5.0.dev0
redis~=4.6.0
openai~=0.27.8
requests~=2.31.0
python-dotenv~=1.0.0
loguru~=0.7.0